Add pinning to ensure that PPA is used
When a package in the main repositories is newer than the one
in the PPA, that one is used. This means that the patches from
the PPA aren't used.

This patch uses pinning to give more priority to the PPA, thus
ensuring that those packages are used, no matter that they are
older than the ones in the main repository.