Allow to connect specific plugs during installation
Sometimes it is needed to be able to connect some plugs during
installation. Although the right way of doing it is by editing
the gadget-desktop snap, this is cumbersome for quick tests.

This patch adds a quick&dirty way of connecting the plugs during
installation. It basically creates a systemd service that is run
at startup and tries to connect the specified plugs. When it
succeeds, it automatically disables itself.

This method must be used only during tests, never for production.