Update XDG folders when changing locale
When the user changes the locale, the XDG folders should be
renamed to their translated names. This is done by
xdg-user-dirs-gtk-update and xdg-user-dirs-update. Unfortunately,
the former was missing, and also some .mo files are required for
them to work.

This patch adds everything.

Unfortunately, there is still some problem: when the folders are
renamed, the desktop icons disappear. After rebooting/restarting
the Gnome Shell session, the icons still don't appear (it seems
to be some problem with nautilus). After a new reboot/restart,
the icons finally do appear.