Merge upstream changes from core22 snap (#62)
* doc: remove reference to using the impish rootfs

As the core22 is now based on the 22.04 rootfs, this TODO can be
removed.

Signed-off-by: Isaac True <isaac.true@canonical.com>

* pam: Add optional faillock

When `/etc/writable/faillock.enabled` is present, then we use
`pam_faillock` which can lock accounts for 900 seconds after 3 wrong
password.

* Remove some unmounting errors on shutdown

Because `/` is a squashfs from `/run/mnt/data`, we cannot unmount
`/run/mnt/data`. (Or `/run/mnt/ubuntu-seed` during install/recovery).

So we need to disable `/run/mnt/data` by adding it to
`shutdown.target`. However we need also to remount read-only. To avoid
with writable bind mounts, we need to do it after `umount.target`.

`/usr/lib/modules` is usually kept busy by systemd-udevd which keeps
running during shutdown. `/usr/lib/modules` is a bind mount of
`/run/mnt/kernel`. So those two mounts also need to be kept alive.

This does not remove errors from finalrd which will be fixed in a
different PR.

Note for testing: this happens after journald has stopped writing to
logs. So you need to use serial port to properly log the errors.

* hooks: Add symlinks for bash completions of snap and snaps

Completion for `snap` must be present in
`/usr/share/bash-completion/completions/` and it must point to the
file in the active Snapd.

Other completion will be added in
`/var/lib/snapd/desktop/bash-completion/completions/` by Snapd. To be
able to access those `XDG_DATA_DIRS` has to be set by Snapd's
`profile.d` file.

This commits also adds snap's `environment.d` file which is used
by user services rather than interactive users.

* Fix mount propagation

To switch root, systemd has to recursively make all mounts private,
then after it recursively make all mounts shared.

However `/run/mnt/*` and `/writable` are used to be bind mounted in
the rest of the file system. For example, there is no reason for mount
`/snap/hello/42` to also show up as
`/writable/system-data/snap/hello/42` and
`/run/mnt/data/system-data/snap/hello/42`.

* Remove polkit policies and rules

Since we do not distribute the polkit daemon, those files are useless.

* Revert "Remove polkit policies and rules"

This reverts commit bb8b56629c87eafe019a2975d65756d5c584b2e1.

* snapd.sshd-keygen.service: fix unit dependency to sshd

The correct name is `ssh.service`

* Rework disabling of unmounts

Some mounts cannot be unmounted during `shutdown.target`.  We have
tried make them part of `shutdown.target` so they are kept alive. But
because they depend on `system-systemd\\x2dfsck.slice` and
`systemd-fsck@.service` which both conflict with `shutdown.target`, it
confuses systemd, and the result of what gets scheduled is not
deterministic.

Instead we remove conflicts of `system-systemd\\x2dfsck.slice` and
`systemd-fsck@.service`. Then `DefaultDependencies=no` on mounts is
enough to disable unmounts.

* static: Add generic nvidia udev rules

These integrate nvidia graphics cards with logind and enable power
management.

Signed-off-by: Dimitri John Ledkov <dimitri.ledkov@canonical.com>

* Fix release branch selection in workflow for pull/push

* Enable 001-extra-packages.test

* hooks: include ppa:ubuntu-security/fde-ice as part of the builds

The security team maintains a PPA with the user-space parts of
cryptsetup support for the Inline Crypto Engine (ICE) code.

This needs to be included in the core build so that FDE on ICE
is supported.

The full ICE implementation uses:
https://github.com/chrisccoulson/linux/tree/dm-blk-crypto
https://github.com/chrisccoulson/cryptsetup/tree/ice-wip

and is getting upstreamed by the kernel and security teams.

* Make all supported locale aliases to C

This is needed for `ln_langinfo(CODESET)` to return the right
encoding.

* static: correct service name used in hctosys rtc fix rule

* static: correct clock epoch file in script fix-hctosys

* snapcraft.yaml: update subiquity version

Signed-off-by: Ondrej Kubik <ondrej.kubik@canonical.com>

* Remove cleanup-timesyncd hack

It always removes the clock stamp so it breaks timesyncd ability to
put forward the clock to a time further than when system was last
powered down, when no rtc is available.

* bootchart: read mode from modeenv instead of kernel command line

Also use a generator to enable it instead of disabling it with
conditions.

* Accept ubuntu_core.bootchart as alias for core.bootchart

To keep consistency with other kernel command line settings.

* get-arg: fix call to printf

Anything with and "A" or a "0" in the name was not parsed correctly
and interpreted as a space instead.

* static/usr/lib: added more deterministic system time to rtc time sync activated by kernel cmdline

* static/usr/lib: remove unused fix-hctosys.service

* static/usr/lib: rtc-sys-time-init@.service must remain after exit

* 90-rtc-sys-time-init.rules: Optimize

* 90-rtc-sys-time-init.rules: do not call systemd-escape

* hooks: enable polkit for unconfined processes

polkit is needed in some cases for communication between services
shipped in the base, for instance when systemd-networkd talks to
systemd-hostnamed to set the hostname. Enable to fix these use cases.
Note that this does not help yet if we want to use it from a snap.

Backported from https://github.com/snapcore/core-base/pull/161

* 66-snapd-autoimport.rules: filter out non filesystem blocks

Because we might run `snap auto-import --mount` on a partition table
block and its partition blocks at the same time, it might cause a
conflict.  We then get an error like `/dev/sda1 already mounted or
mount point busy.`.

To go around this issue, we trigger the service only for devices that
are identified as filesystems.

* snapd.autoimport-device@.service: do not run along snapd.autoimport.service

There is a conflict between snapd.autoimport.service and
snapd.autoimport-device@.service. This causes error messages because
both cannot mount in the same time. So they cannot be run at the same
time.

* run-snapd-from-snap: set the current symlink before bootstrapping

* hooks: create journald drop-in directory

* snapcraft.yaml: move or remove extra files from plymouth theme

Fixes #205

* bump subiquity and probert

* hooks: remove duplication of apps-bin-path.sh symlink in 022-setup-path.chroot

* hooks: remove 009-locale-archive.chroot script, as we ship locales-all

---------

Signed-off-by: Isaac True <isaac.true@canonical.com>
Signed-off-by: Dimitri John Ledkov <dimitri.ledkov@canonical.com>
Signed-off-by: Ondrej Kubik <ondrej.kubik@canonical.com>
Co-authored-by: Isaac True <isaac.true@canonical.com>
Co-authored-by: Valentin David <valentin.david@canonical.com>
Co-authored-by: Michael Vogt <mvo@ubuntu.com>
Co-authored-by: alfonsosanchezbeato <alfonso.sanchez-beato@canonical.com>
Co-authored-by: Valentin David <me@valentindavid.com>
Co-authored-by: Dimitri John Ledkov <dimitri.ledkov@canonical.com>
Co-authored-by: Ernest Lotter <ernest.lotter@canonical.com>
Co-authored-by: Ondrej Kubik <ondrej.kubik@canonical.com>
Co-authored-by: Philip Meulengracht <the_meulengracht@hotmail.com>
Co-authored-by: Michael Hudson-Doyle <michael.hudson@canonical.com>